[TOW-61] 엔터티 변경
* 엔터티 이름 및 변수명 변견
