+++ TodayWod/Model/Wod/ProgramModel.swift
@@ -7,50 +7,50 @@
 
 import Foundation
 
-struct ProgramsModel: Equatable {
+struct ProgramModel: Equatable {
 
     let id: UUID
     let methodType: ProgramMethodType
     let level: LevelType
-    let weeklyWorkouts: [WeeklyWorkoutModel]
+    let dayWorkouts: [DayWorkoutModel]
 
-    init(entity: ProgramsEntity) {
+    init(entity: ProgramEntity) {
         self.id = entity.id
         self.methodType = ProgramMethodType(rawValue: entity.methodType) ?? .body
         self.level = LevelType(rawValue: entity.level) ?? .beginner
-        self.weeklyWorkouts = entity.weeklyWorkouts.map {
-            WeeklyWorkoutModel(entity: $0 as! WeeklyWorkoutEntity)
+        self.dayWorkouts = entity.dayWorkouts.map {
+            DayWorkoutModel(entity: $0 as! DayWorkoutEntity)
         }
     }
 
-    init(id: UUID, methodType: ProgramMethodType, level: LevelType, weeklyWorkoutPrograms: [WeeklyWorkoutModel]) {
+    init(id: UUID, methodType: ProgramMethodType, level: LevelType, dayWorkouts: [DayWorkoutModel]) {
         self.id = id
         self.methodType = methodType
         self.level = level
-        self.weeklyWorkouts = weeklyWorkoutPrograms
+        self.dayWorkouts = dayWorkouts
     }
 
 }
 
-extension ProgramsModel {
+extension ProgramModel {
 
-    static let mock: Self = .init(id: UUID(), methodType: .body, level: .beginner, weeklyWorkoutPrograms: [
+    static let mock: Self = .init(id: UUID(), methodType: .body, level: .beginner, dayWorkouts: [
         .init(id: UUID(),
               type: .start,
               title: "알파 데이",
               subTitle: "한 주를 힘차게 시작하는",
               expectedMinutes: 60,
-              minExpectedCalories: 445,
-              maxExpectedCalories: 550,
+              minExpectedCalorie: 445,
+              maxExpectedCalorie: 550,
               workoutInfos: [
-                .init(id: UUID(), type: .warmUp, workOutItems: [
+                .init(id: UUID(), type: .warmUp, wods: [
                     .init(id: UUID(),
                           title: "트레드밀",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 1,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false)
                           ]),
                     .init(id: UUID(),
@@ -59,19 +59,19 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .main, workOutItems: [
+                .init(id: UUID(), type: .main, wods: [
                     .init(id: UUID(),
                           title: "체스트 프레스 머신",
                           subTitle: "lowing abc",
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -82,7 +82,7 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -93,20 +93,20 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 30,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .coolDown, workOutItems: [
+                .init(id: UUID(), type: .coolDown, wods: [
                     .init(id: UUID(),
                           title: "엘립티컬 머신",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -116,7 +116,7 @@ extension ProgramsModel {
                           unit: .seconds,
                           unitValue: 20,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -127,17 +127,17 @@ extension ProgramsModel {
               title: "타이탄 데이",
               subTitle: "한 주를 힘차게 시작하는",
               expectedMinutes: 60,
-              minExpectedCalories: 445,
-              maxExpectedCalories: 550,
+              minExpectedCalorie: 445,
+              maxExpectedCalorie: 550,
               workoutInfos: [
-                .init(id: UUID(), type: .warmUp, workOutItems: [
+                .init(id: UUID(), type: .warmUp, wods: [
                     .init(id: UUID(),
                           title: "트레드밀",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 1,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false)
                           ]),
                     .init(id: UUID(),
@@ -146,19 +146,19 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .main, workOutItems: [
+                .init(id: UUID(), type: .main, wods: [
                     .init(id: UUID(),
                           title: "체스트 프레스 머신",
                           subTitle: "lowing abc",
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -169,7 +169,7 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -180,20 +180,20 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 30,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .coolDown, workOutItems: [
+                .init(id: UUID(), type: .coolDown, wods: [
                     .init(id: UUID(),
                           title: "엘립티컬 머신",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -203,7 +203,7 @@ extension ProgramsModel {
                           unit: .seconds,
                           unitValue: 20,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -214,17 +214,17 @@ extension ProgramsModel {
               title: "히어로 데이",
               subTitle: "한 주를 힘차게 시작하는",
               expectedMinutes: 60,
-              minExpectedCalories: 445,
-              maxExpectedCalories: 550,
+              minExpectedCalorie: 445,
+              maxExpectedCalorie: 550,
               workoutInfos: [
-                .init(id: UUID(), type: .warmUp, workOutItems: [
+                .init(id: UUID(), type: .warmUp, wods: [
                     .init(id: UUID(),
                           title: "트레드밀",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 1,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false)
                           ]),
                     .init(id: UUID(),
@@ -233,19 +233,19 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .main, workOutItems: [
+                .init(id: UUID(), type: .main, wods: [
                     .init(id: UUID(),
                           title: "체스트 프레스 머신",
                           subTitle: "lowing abc",
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -256,7 +256,7 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -267,20 +267,20 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 30,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .coolDown, workOutItems: [
+                .init(id: UUID(), type: .coolDown, wods: [
                     .init(id: UUID(),
                           title: "엘립티컬 머신",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -290,7 +290,7 @@ extension ProgramsModel {
                           unit: .seconds,
                           unitValue: 20,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -301,17 +301,17 @@ extension ProgramsModel {
               title: "블레디즈 데이",
               subTitle: "한 주를 힘차게 시작하는",
               expectedMinutes: 60,
-              minExpectedCalories: 445,
-              maxExpectedCalories: 550,
+              minExpectedCalorie: 445,
+              maxExpectedCalorie: 550,
               workoutInfos: [
-                .init(id: UUID(), type: .warmUp, workOutItems: [
+                .init(id: UUID(), type: .warmUp, wods: [
                     .init(id: UUID(),
                           title: "트레드밀",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 1,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false)
                           ]),
                     .init(id: UUID(),
@@ -320,19 +320,19 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .main, workOutItems: [
+                .init(id: UUID(), type: .main, wods: [
                     .init(id: UUID(),
                           title: "체스트 프레스 머신",
                           subTitle: "lowing abc",
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -343,7 +343,7 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -354,20 +354,20 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 30,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .coolDown, workOutItems: [
+                .init(id: UUID(), type: .coolDown, wods: [
                     .init(id: UUID(),
                           title: "엘립티컬 머신",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -377,7 +377,7 @@ extension ProgramsModel {
                           unit: .seconds,
                           unitValue: 20,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -388,17 +388,17 @@ extension ProgramsModel {
               title: "피닉스 데이",
               subTitle: "한 주를 힘차게 시작하는",
               expectedMinutes: 60,
-              minExpectedCalories: 445,
-              maxExpectedCalories: 550,
+              minExpectedCalorie: 445,
+              maxExpectedCalorie: 550,
               workoutInfos: [
-                .init(id: UUID(), type: .warmUp, workOutItems: [
+                .init(id: UUID(), type: .warmUp, wods: [
                     .init(id: UUID(),
                           title: "트레드밀",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 1,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false)
                           ]),
                     .init(id: UUID(),
@@ -407,19 +407,19 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .main, workOutItems: [
+                .init(id: UUID(), type: .main, wods: [
                     .init(id: UUID(),
                           title: "체스트 프레스 머신",
                           subTitle: "lowing abc",
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -430,7 +430,7 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -441,20 +441,20 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 30,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .coolDown, workOutItems: [
+                .init(id: UUID(), type: .coolDown, wods: [
                     .init(id: UUID(),
                           title: "엘립티컬 머신",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -464,7 +464,7 @@ extension ProgramsModel {
                           unit: .seconds,
                           unitValue: 20,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -475,17 +475,17 @@ extension ProgramsModel {
               title: "에이펙스 데이",
               subTitle: "한 주를 힘차게 시작하는",
               expectedMinutes: 60,
-              minExpectedCalories: 445,
-              maxExpectedCalories: 550,
+              minExpectedCalorie: 445,
+              maxExpectedCalorie: 550,
               workoutInfos: [
-                .init(id: UUID(), type: .warmUp, workOutItems: [
+                .init(id: UUID(), type: .warmUp, wods: [
                     .init(id: UUID(),
                           title: "트레드밀",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 1,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false)
                           ]),
                     .init(id: UUID(),
@@ -494,19 +494,19 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .main, workOutItems: [
+                .init(id: UUID(), type: .main, wods: [
                     .init(id: UUID(),
                           title: "체스트 프레스 머신",
                           subTitle: "lowing abc",
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -517,7 +517,7 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 12,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
@@ -528,20 +528,20 @@ extension ProgramsModel {
                           unit: .repetitions,
                           unitValue: 30,
                           set: 3,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false),
                             .init(id: UUID(), order: 3, unitValue: 3, isCompleted: false)
                           ]),
                 ]),
-                .init(id: UUID(), type: .coolDown, workOutItems: [
+                .init(id: UUID(), type: .coolDown, wods: [
                     .init(id: UUID(),
                           title: "엘립티컬 머신",
                           subTitle: "lowing abc",
                           unit: .minutes,
                           unitValue: 5,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
@@ -551,7 +551,7 @@ extension ProgramsModel {
                           unit: .seconds,
                           unitValue: 20,
                           set: 2,
-                          wodSet: [
+                          wodSets: [
                             .init(id: UUID(), order: 1, unitValue: 1, isCompleted: false),
                             .init(id: UUID(), order: 2, unitValue: 2, isCompleted: false)
                           ]),
