+++ TodayWod/Model/Wod/WodModel.swift
@@ -15,7 +15,7 @@ struct WodModel: Equatable, Identifiable {
     let unit: ExerciseUnit
     let unitValue: Int
     let set: Int
-    var wodSet: [WodSetModel]
+    var wodSets: [WodSetModel]
 
     init(entity: WodEntity) {
         self.id = entity.id
@@ -24,17 +24,17 @@ struct WodModel: Equatable, Identifiable {
         self.unit = ExerciseUnit(rawValue: entity.unit) ?? .seconds
         self.unitValue = Int(entity.unitValue)
         self.set = Int(entity.set)
-        self.wodSet = entity.wodSet.map { WodSetModel(entity: $0 as! WodSetEntity) }
+        self.wodSets = entity.wodSets.map { WodSetModel(entity: $0 as! WodSetEntity) }
     }
 
-    init(id: UUID, title: String, subTitle: String, unit: ExerciseUnit, unitValue: Int, set: Int, wodSet: [WodSetModel]) {
+    init(id: UUID, title: String, subTitle: String, unit: ExerciseUnit, unitValue: Int, set: Int, wodSets: [WodSetModel]) {
         self.id = id
         self.title = title
         self.subTitle = subTitle
         self.unit = unit
         self.unitValue = unitValue
         self.set = set
-        self.wodSet = wodSet
+        self.wodSets = wodSets
     }
 
 }
