+++ TodayWod/TodayWodCoreData.xcdatamodeld/TodayWodCoreData.xcdatamodel/contents
@@ -1,24 +1,19 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="23231" systemVersion="24A335" minimumToolsVersion="Automatic" sourceLanguage="Swift" usedWithSwiftData="YES" userDefinedModelVersionIdentifier="">
     <entity name="DayWorkoutEntity" representedClassName="DayWorkoutEntity" syncable="YES">
-        <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
-        <attribute name="type" attributeType="String"/>
-        <relationship name="wods" optional="YES" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="WodEntity"/>
-    </entity>
-    <entity name="ProgramsEntity" representedClassName="ProgramsEntity" syncable="YES">
-        <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
-        <attribute name="level" attributeType="String"/>
-        <attribute name="methodType" attributeType="String"/>
-        <relationship name="weeklyWorkouts" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="WeeklyWorkoutEntity"/>
-    </entity>
-    <entity name="WeeklyWorkoutEntity" representedClassName="WeeklyWorkoutEntity" syncable="YES">
         <attribute name="expectedMinutes" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
-        <attribute name="maxExpectedCalories" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="minExpectedCalories" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
+        <attribute name="maxExpectedCalorie" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
+        <attribute name="minExpectedCalorie" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="subTitle" attributeType="String"/>
         <attribute name="title" attributeType="String"/>
         <attribute name="type" attributeType="String"/>
+        <relationship name="workouts" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="WorkoutEntity"/>
+    </entity>
+    <entity name="ProgramEntity" representedClassName="ProgramEntity" syncable="YES">
+        <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
+        <attribute name="level" attributeType="String"/>
+        <attribute name="methodType" attributeType="String"/>
         <relationship name="dayWorkouts" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="DayWorkoutEntity"/>
     </entity>
     <entity name="WodEntity" representedClassName="WodEntity" syncable="YES">
@@ -28,12 +23,17 @@
         <attribute name="title" attributeType="String"/>
         <attribute name="unit" attributeType="String"/>
         <attribute name="unitValue" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
-        <relationship name="wodSet" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="WodSetEntity"/>
+        <relationship name="wodSets" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="WodSetEntity"/>
     </entity>
     <entity name="WodSetEntity" representedClassName="WodSetEntity" syncable="YES">
         <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
         <attribute name="isCompleted" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="order" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="unitValue" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
+    </entity>
+    <entity name="WorkoutEntity" representedClassName="WorkoutEntity" syncable="YES">
+        <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
+        <attribute name="type" attributeType="String"/>
+        <relationship name="wods" optional="YES" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="WodEntity"/>
     </entity>
 </model>
